add basic layout for Knowledge
