+++ src/routes/sessions/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Empty from '../../lib/components/EmptyMessage.svelte';
+	import Empty from '$lib/components/EmptyMessage.svelte';
 </script>
 
 <Empty>Create a new session or choose an existing one from the list</Empty>
